pyproject.toml
[project]
name = "cozy"
version = "0.1.0"
description = "A simple API wrapper to generate images using ComfyUI"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.118.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.37.0",
    "websocket-client>=1.8.0",
]
